Use downcast instead of extract where possible
[docs](https://pyo3.rs/v0.21.1/performance#extract-versus-downcast) suggest
this is faster, and can lead to better error messages in the
`bloom.update(bloom)` case: before, `extract::<PyRef<Bloom>>()` would fail, so
we would fall back to trying (and failing) to iterate the Bloom object. Now,
instead, we get a clear error that `bloom` is already borrowed mutably.

We could instead take `&Bound<'_, Bloom>`, and check `slf.is(other)` before
borrowing, to avoid the error entirely, but it makes the signature and body of
the functions a fair bit uglier, not sure if it's worth it.